user can log out
